clean up / refactor Seamless inputs
* no need for withSeamlessness HoC - Input not used, TextArea now has outline
* rename SeamlessTextArea -> WrappingInput (it behaves like an input, despite being a textarea)
