+++ src/components/Forms/withChangeHandler.js
@@ -23,6 +23,7 @@ const withChangeHandler = WrappedComponent => {
 
     render() {
       const { value, className, ...otherProps } = this.props;
+
       return (
         <WrappedComponent
           {...otherProps}
