--- src/components/SeamlessInput/SeamlessInput.js
@@ -1,8 +1,0 @@
-import React from "react";
-import withSeamlessness from "components/EditorSurface/withSeamlessness";
-import withChangeHandler from "components/Forms/withChangeHandler";
-import { flow } from "lodash";
-
-const Input = props => <input type="text" {...props} />;
-
-export default flow(withChangeHandler, withSeamlessness)(Input);
